improved naming - tabs and tables in ui now correspond
