+++ lca_activity_browser/app/ui/tables/__init__.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
-from .activity import ActivitiesTable   # ActivitiesTableNew
-from .biosphere import BiosphereFlowsTable
+from .inventory import ActivitiesTable   # ActivitiesTableNew
+from .inventory import BiosphereFlowsTable
 from .calculation_setups import (
     CSActivityTable,
     CSList,
     CSMethodsTable,
 )
-from .database import DatabasesTable
-from .exchange import ExchangeTable
+from .inventory import DatabasesTable
+from .activity import ExchangeTable
 from .history import ActivitiesHistoryTable
 from .ia import CFTable, MethodsTable
 from .lca_results import LCAResultsTable
